requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
ConfigArgParse==1.2.1
Flask==1.1.2
gevent==1.5a3
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
locust==0.0
locustio==0.14.5
MarkupSafe==1.1.1
msgpack==1.0.0
pkg-resources==0.0.0
pprint==0.1
psutil==5.7.0
python-dotenv==0.12.0
pyzmq==19.0.0
requests==2.23.0
rocketchat-API==1.3.1
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
